chore(VBadge): move extract function to helpers
