+++ packages/vuetify/src/util/helpers.ts
@@ -165,6 +165,21 @@ export const keyCodes = Object.freeze({
   pagedown: 34,
 })
 
+export function extract (obj: Dictionary<unknown>, properties: string[]) {
+  const extracted: Dictionary<unknown> = {}
+  const rest: Dictionary<unknown> = {}
+
+  Object.entries(obj).forEach(([key, value]) => {
+    if (properties.includes(key)) {
+      extracted[key] = value
+    } else {
+      rest[key] = value
+    }
+  })
+
+  return [extracted, rest]
+}
+
 export function keys<O> (o: O) {
   return Object.keys(o) as (keyof O)[]
 }
